requirements.txt
playwright==1.22.0
pydantic==1.9.1
PyYAML==6.0
